+++ workspace/src/barc/src/plotCarTrajectory.py
@@ -207,7 +207,7 @@ def _initializeFigure_xy(map, mode):
     plt.ion()
     axtr = plt.axes()
 
-    Points = np.floor(10 * (map.PointAndTangent[-1, 3] + map.PointAndTangent[-1, 4]))
+    Points = int(np.floor(10 * (map.PointAndTangent[-1, 3] + map.PointAndTangent[-1, 4])))
     Points1 = np.zeros((Points, 2))
     Points2 = np.zeros((Points, 2))
     Points0 = np.zeros((Points, 2))
